test(openai): separate reasoning replay coverage
